reorganised file structure for view-roles and added the index file
