+++ src/app/view-roles/components/index .tsx
@@ -1,0 +1,4 @@
+import ViewCard from './card/viewcard';
+import DeleteCard from './card/deletecard';
+import ViewTable from './table/viewtable';
+export { ViewCard, DeleteCard, ViewTable };
